rename author folder
