+++ authors/author_manager.rb
